consolidate waveform code in data module
