--- projects/data/data/timeslide_waveforms/__init__.py
@@ -1,1 +1,0 @@
-from .timeslide_waveforms import main, parser
